+++ src/main/java/com/avanzada/alojamientos/DTO/ProfileHostDTO.java
@@ -3,8 +3,8 @@ package com.avanzada.alojamientos.DTO;
 import java.time.LocalDateTime;
 
 public record ProfileHostDTO(
-        Integer id,
-        Integer hostId,
+        long id,
+        long hostId,
         String businessName,
         Boolean verified,
         LocalDateTime createdAt
